+++ th2_data_services/data_source/lwdp/streams.py
@@ -48,6 +48,16 @@ class Stream:
             return f"&stream={self._alias}"
         return f"&stream={self._alias}:{self._direction}"
 
+    def convert_to_dict_format(self) -> dict:
+        direction_mapper = {
+            "1": ["IN"],
+            "2": ["OUT"],
+        }
+        result = {"sessionAlias": self._alias}
+        if self._direction:
+            result["direction"] = direction_mapper[self._direction]
+        return result
+
 
 class Streams:
     """General interface for composite streams of lwdp ds.
@@ -92,3 +102,29 @@ class Streams:
         if self._direction is None:
             return "&".join([f"stream={alias}" for alias in self._aliases])
         return "&".join([f"stream={stream}:{self._direction}" for stream in self._aliases])
+
+    def convert_to_dict_format(self) -> List[dict]:
+        return [
+            Stream(stream, self._direction).convert_to_dict_format() for stream in self._aliases
+        ]
+
+
+def _convert_stream_to_dict_format(streams):
+    def map_func(s):
+        if isinstance(s, Stream):
+            return [s.convert_to_dict_format()]
+
+        if isinstance(s, Streams):
+            return s.convert_to_dict_format()
+
+        if isinstance(s, str):
+            if ":" in s:
+                stream_obj = Stream(*s.split(":"))
+            else:
+                stream_obj = Stream(s)
+            return [stream_obj.convert_to_dict_format()]
+        return [s]
+
+    if isinstance(streams, List):
+        return [item for sublist in map(map_func, streams) for item in sublist]
+    return map_func(streams)