+++ th2_data_services/data_source/lwdp/utils/iter_status_manager.py
@@ -34,4 +34,8 @@ class StatusUpdateManager:
         self.__data = data
 
     def update(self, status):
-        self.__data.update_metadata({"Iter statuses": IterStatus(**status)})
+        try:
+            self.__data.update_metadata({"Iter statuses": IterStatus(**status)})
+        except:
+            print(f"Unable to create an IterStatus object with these keys: {status}")
+            raise