+++ release_notes.md
@@ -256,10 +256,22 @@ BugFixes without ticket
    its value is either `sse` or `json` and will dictate which method will be used to fetch messages. Added GetMessagesByBookByGroupsJson, 
    GetMessagesByPageByGroupsJson commands, they return messages in real time. Now, for example, GetMessagesByBookByGroups
    will use GetMessagesByBookByGroupsSse or GetMessagesByBookByGroupsJson depending on `request_mode` (default value is `sse`).
+2. [TH2-5218] Added `DownloadEventsByBookByScopeGzip` and `GetEventsByBookByScopeJson` commands.
 
 ## Improvements
 1. [TH2-5228] Session objects are now used to make requests.
 
 ## BugFixes
 1. [TH2-5222] - Fix `DownloadMessagesByPageGzip`, it constructed `DownloadMessagesByPageByGroupsGzip` incorrectly.
-2. [TH2-5243] - Fix backward compatibility issues with `GetMessagesByBookByGroups` and `GetMessagesByPageByGroups`.
+2. [TH2-5243] - Fix backward compatibility issues with `GetMessagesByBookByGroups` and `GetMessagesByPageByGroups`,
+   added 'limit' and 'search_direction' parameters to 'post_download_messages'.
+
+# v3.1.2.0
+
+## Features
+1. [TH2-5258] Added `gzip: bool` parameter that indicates whether to include or not `gzip` in the Accept-Encoding header.
+
+## Improvements
+1. [TH2-5260] More details are provided in case an exception occurs in `StatusUpdateManager.update()`.
+2. [TH2-5263] Retries are added to `IHTTPDataSource.check_connect()`.
+3. [TH2-5264] Retries are added to `nest_asyncio.apply()`.